+++ node/src/chain.rs
@@ -34,6 +34,8 @@ use tracing::{debug, error, info, warn};
 
 use self::acceptor::Acceptor;
 use self::fsm::SimpleFSM;
+#[cfg(feature = "archive")]
+use crate::archive::Archive;
 use crate::database::rocksdb::MD_HASH_KEY;
 use crate::database::{Ledger, Metadata};
 use crate::{database, vm, LongLivedService, Message, Network};
@@ -59,6 +61,8 @@ pub struct ChainSrv<N: Network, DB: database::DB, VM: vm::VMExecution> {
     event_sender: Sender<Event>,
     genesis_timestamp: u64,
     dusk_key: BlsPublicKey,
+    #[cfg(feature = "archive")]
+    archive: Archive,
 }
 
 #[async_trait]
@@ -89,6 +93,8 @@ impl<N: Network, DB: database::DB, VM: vm::VMExecution>
             db,
             network,
             vm,
+            #[cfg(feature = "archive")]
+            self.archive.clone(),
             self.max_consensus_queue_size,
             self.event_sender.clone(),
             self.dusk_key,
@@ -251,6 +257,7 @@ impl<N: Network, DB: database::DB, VM: vm::VMExecution> ChainSrv<N, DB, VM> {
         event_sender: Sender<Event>,
         genesis_timestamp: u64,
         dusk_key: BlsPublicKey,
+        #[cfg(feature = "archive")] archive: Archive,
     ) -> Self {
         info!(
             "ChainSrv::new with keys_path: {}, max_inbound_size: {}",
@@ -265,6 +272,8 @@ impl<N: Network, DB: database::DB, VM: vm::VMExecution> ChainSrv<N, DB, VM> {
             event_sender,
             genesis_timestamp,
             dusk_key,
+            #[cfg(feature = "archive")]
+            archive,
         }
     }